refactor: switch to a chain/network based file structure
Considering the upcoming network packages, using the current file
structure doesn't fullfil our needs anymore. If we continue using it, we
have to put both network and chain packages in the same directory, which
is not arranged enough.

Add two new directories inside `packages`, one for networks and one for
chains. In addition, apply required changes to all configration files.