chore(abstract-chain): add new params to `getBoxesByTokenId`
Add `address`, `offset` and `limit` params to `getBoxesByTokenId`. The
first one is added because there is no node api which returns all boxes
by token id and so, without the param, the ergo network package cannot
be implemented. The latter two params are added to keep consistency of
apis.

In addition, apply required changes to `ErgoChain` and upgrade its
package by a patch version.